+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version 1.2.0
 - Added property *physical_sector_size* to *Device* object
 - Added "sudo" property to the Smartctl class for POSIX systems.
 - Added global SMARTCTL object used for defaults.
+- Added sector sizes to diagnostics class and other minor improvements to its structure.
 - Minor changes on parsing regexes
 - Checks for temperature units (farenheit) and converts to celsius
 - Fixed MacOS compatibility issues up to version 11.x